Várias atualizações: conversão de unidades, SQL para tipos de quantidade e melhorias na interface

+++ sql/recreate_movements_table.sql
@@ -6,7 +6,7 @@ CREATE TABLE movements (
   id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
   product_id UUID NOT NULL REFERENCES products(id) ON DELETE CASCADE,
   type TEXT CHECK (type IN ('entrada', 'saida')) NOT NULL,
-  quantity INTEGER NOT NULL,
+  quantity DECIMAL(10,4) NOT NULL,
   user_id UUID REFERENCES auth.users(id),
   notes TEXT,
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
